% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpledice.R
\name{simpleDice}
\alias{simpleDice}
\title{Give a random number or letter}
\usage{
simpleDice(n, sides = "numbers")
}
\arguments{
\item{n}{integer}

\item{sides}{string "letters" or "nambers}
}
\value{
integer or character
}
\description{
Get a random number (by default) or letter
from a given range
}
\examples{
simpleDice(50, "letters")
}
